fix: comprehensive environment variable detection improvements
- Enhanced Docusaurus configuration source detection with multiple fallbacks
- Improved webpack DefinePlugin variable access patterns
- Added comprehensive debugging for environment variable resolution
- Fixed window scope access for webpack-injected variables
- Added fallback patterns for customFields access
- Enhanced error reporting for missing environment variables

This should resolve the issue where environment variables are not being
properly detected in the browser, causing the frontend to fall back to
production API instead of using branch-specific API URLs.
